Separated a map as Map.js and wrote a foundation code for doctors components
