+++ .env
@@ -1,7 +1,2 @@
-<<<<<<< HEAD
-REACT_APP_API=https://findmeadoc.herokuapp.com/
-=======
 REACT_APP_API=https://findmeadoc.herokuapp.com
->>>>>>> main
-
-
+BACKEND_CONNECTION=https://findmeadoc.herokuapp.com
