+++ src/components/doctors/DoctorShow.js
@@ -17,21 +17,28 @@ function DoctorShow() {
         <div key={doctor._id}>
           <h1>This is a DoctorShow page</h1>
           <h4>{doctor.fullName}</h4>
-          {doctor.specialties.map((element) => (
-            <p key={element}>{element}</p>
-          ))}
-          {doctor.languages.map((element) => (
-            <p key={element}>{element}</p>
-          ))}
-          {/* {doctor.address.map((element, i) => (
-            // console.log(Object.values(element))
-            <div key={i}>
-              <p key={Object.values(element)[0]}>{Object.values(element)[0]}</p>
-              <p key={Object.values(element)[2]}>{Object.values(element)[2]}</p>
-              <p key={Object.values(element)[3]}>{Object.values(element)[3]}</p>
-              <p key={Object.values(element)[4]}>{Object.values(element)[4]}</p>
-            </div>
-          ))} */}
+          <p>Specialities</p>
+          <ul>
+            {doctor.specialties.map((element, i) => (
+              <li key={i}>{element}</li>
+            ))}
+          </ul>
+          <p>Languages</p>
+          <ul>
+            {doctor.languages.map((element, i) => (
+              <li key={i}>{element}</li>
+            ))}
+          </ul>
+          <p>Address</p>
+          <ul>
+            {doctor.address.addressLine2 && (
+              <li>{doctor.address.addressLine2}</li>
+            )}
+            <li>{doctor.address.addressLine1}</li>
+            <li>{doctor.address.town}</li>
+            <li>{doctor.address.country}</li>
+            <li>{doctor.address.postcode.toUpperCase()}</li>
+          </ul>
         </div>
       ) : (
         <p>Loading...</p>
